Bugfix and enhance "setup_python.py"
Bugfix fallback in `get_best_variant()`

Check if local python executable is in PATH and warn if not.
Don't download if again, if it's not in PATH.

Check if "zstd" or "gzip" is installed for decompression and prefer "zstd" over "gzip".

Handle different directory structures in archive: with and without `install` directory.